refactor load of icons & bug while edit dimension
